+++ packages/core/src/create-extension.js
@@ -18,6 +18,10 @@ const defaultOptions = {
 const createExtension = (receivedName, receivedOptions = {}) => {
   const { hooks, stack } = getState();
 
+  if (!receivedName) {
+    throw new Error(`createExtension() missing Extension name!`);
+  }
+
   const name =
     receivedName.substr(0, 1) === '$'
       ? getTarget(receivedName.substr(1))
@@ -34,6 +38,8 @@ const createExtension = (receivedName, receivedOptions = {}) => {
     ...receivedOptions,
   };
 
+  // console.log('@@@@', name);
+
   const actions = (hooks[name] || []).filter((h) => h.enabled === true);
 
   const writeLog = () => {
@@ -46,6 +52,7 @@ const createExtension = (receivedName, receivedOptions = {}) => {
 
   if (options.mode === 'parallel') {
     return new Promise(async (resolve, reject) => {
+      // console.log('@@@@ PARALLEL', name);
       try {
         const promises = actions.map((action) => runAction(action, options));
         const results = await Promise.all(promises);
@@ -66,9 +73,11 @@ const createExtension = (receivedName, receivedOptions = {}) => {
 
   if (options.mode === 'serie') {
     return new Promise(async (resolve, reject) => {
+      // console.log('@@@@ SERIE', name);
       try {
         const results = [];
         for (const action of actions) {
+          // console.log('>>>', action);
           results.push(await runAction(action, options));
         }
         writeLog();
@@ -88,6 +97,7 @@ const createExtension = (receivedName, receivedOptions = {}) => {
 
   // Edit the value of the args and return it for further iteration
   if (options.mode === 'waterfall') {
+    // console.log('@@@@ WATERFALL', name);
     const results = [];
     let args = options.args;
 
@@ -105,6 +115,7 @@ const createExtension = (receivedName, receivedOptions = {}) => {
 
   // synchronous execution with arguments
   try {
+    // console.log('@@@@ SYNC', name);
     const results = actions.map((action) => runActionSync(action, options));
     writeLog();
     pullStack();