+++ packages/kitchensink/service-postgres/service-postgres.test.js
@@ -1,0 +1,5 @@
+describe('service-postgres-standalone', () => {
+  it('should work', () => {
+    console.log('it dows');
+  });
+});